split mongo to other project
