+++ src/EventLite.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "POC.Documents", "POC.Docume
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EventLite", "EventLite\EventLite.csproj", "{1F8DA2CE-9EB1-46E4-9B65-B658866F78F8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventLite.MongoDB", "databases\EventLite.MongoDB\EventLite.MongoDB.csproj", "{9292DD24-FEBE-4154-B1E9-702E3251C7DE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{1F8DA2CE-9EB1-46E4-9B65-B658866F78F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1F8DA2CE-9EB1-46E4-9B65-B658866F78F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1F8DA2CE-9EB1-46E4-9B65-B658866F78F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9292DD24-FEBE-4154-B1E9-702E3251C7DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9292DD24-FEBE-4154-B1E9-702E3251C7DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9292DD24-FEBE-4154-B1E9-702E3251C7DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9292DD24-FEBE-4154-B1E9-702E3251C7DE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
