--- src/EventLite/Streams/DTO/CommitDTO.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace EventLite.Streams.DTO
-{
-    internal class CommitDTO
-    {
-        public Guid StreamId { get; set; }
-        public int CommitNumber { get; set; }
-        public long Timestamp { get; set; }
-        public object Event { get; set; }
-    }
-}
