--- src/EventLite/Streams/DTO/EventStreamDTO.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace EventLite.Streams.DTO
-{
-    internal class EventStreamDTO
-    {
-        public Guid StreamId { get; set; }
-        public int HeadRevision { get; set; }
-        public int UnsnapshottedCommits { get; set; }
-        public int SnapshotRevision { get; set; }
-    }
-}
