--- src/EventLite/Streams/StreamManager/Implementation/MongoDB/MongoDBSettings.cs
@@ -1,8 +1,0 @@
-﻿namespace EventLite.Streams.StreamManager.Implementation.MongoDB
-{
-    public class MongoDBSettings
-    {
-        public string ConnectionString { get; set; }
-        public string Database { get; set; }
-    }
-}
