+++ src/POC.Documents/IoC.cs
@@ -1,27 +1,29 @@
 ﻿using System;
+using EventLite;
+using EventLite.MongoDB;
 using Microsoft.Extensions.DependencyInjection;
 using POC.Documents.Events;
-using EventLite;
-using EventLite.Streams.StreamManager.Implementation.MongoDB;
 
 namespace POC.Documents
 {
     public static class IoC
     {
         private static IServiceCollection _services { get; set; }
+
         public static IServiceProvider ServiceProvider => _services.BuildServiceProvider();
 
         public static void RegisterServices()
         {
             _services = new ServiceCollection();
 
-            _services.EventSourcingPersistenceWithMongo(new MongoDBSettings
+            _services.PersistWithMongoDB(new MongoDBSettings
             {
                 ConnectionString = "mongodb://localhost",
                 Database = "documents_stream"
             });
 
-            _services.RegisterEventSourcing(new AggregateSettings {
+            _services.RegisterEventSourcing(new AggregateSettings
+            {
                 CommitsBeforeSnapshot = 5
             });
