+++ src/POC.Documents/POC.Documents.csproj
@@ -10,7 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\EventLite\EventLite.csproj" />
+    <ProjectReference Include="..\databases\EventLite.MongoDB\EventLite.MongoDB.csproj" />
   </ItemGroup>
 
 </Project>
