+++ src/databases/EventLite.MongoDB/DTO/DtoProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using EventLite.Streams;
+
+namespace EventLite.MongoDB.DTO
+{
+    internal class DtoProfile : Profile
+    {
+        public DtoProfile()
+        {
+            CreateMap<Commit, CommitDTO>().ReverseMap();
+            CreateMap<EventStream, EventStreamDTO>().ReverseMap();
+            CreateMap<Snapshot<object>, SnapshotDTO>().ReverseMap();
+        }
+    }
+}
