+++ src/databases/EventLite.MongoDB/DTO/EventStreamDTO.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace EventLite.MongoDB.DTO
+{
+    internal class EventStreamDTO
+    {
+        public Guid StreamId { get; set; }
+
+        public int HeadRevision { get; set; }
+
+        public int UnsnapshottedCommits { get; set; }
+
+        public int SnapshotRevision { get; set; }
+    }
+}
