+++ src/databases/EventLite.MongoDB/DTO/SnapshotDTO.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace EventLite.MongoDB.DTO
+{
+    internal class SnapshotDTO
+    {
+        public Guid StreamId { get; set; }
+
+        public int SnapshotRevision { get; set; }
+
+        public int SnapshotHeadCommit { get; set; }
+
+        public object SnapshotData { get; set; }
+    }
+}
