+++ src/databases/EventLite.MongoDB/MongoDBSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace EventLite.MongoDB
+{
+    public class MongoDBSettings
+    {
+        public string ConnectionString { get; set; }
+
+        public string Database { get; set; }
+    }
+}
